move edit routing logic into container and add prop, move mock user into container
